* 460-study-mgmt-page (#462)

* Started feature 460-study-mgmt-page

* Study description and logo

* Add tests

* Update seed.ts

* 496-gen-id (#502)

* Started feature 496-gen-id

* Add participant IDs

* Fix tests

* 130-tcs (#501)

* Started feature 130-tcs

* Terms and conditions box

* Fix tests

* 492-log-recalc (#500)

* Started feature 492-log-recalc

* Log answer recalc

* Fix test

* 493-otp (#494)

* Started feature 493-otp

* OTP and password retries

* complete unfinished test

* Fix test

* 487-multi-study-families (#488)

* Started feature 487-multi-study-families

* Update family editing

* Add tests

* docs, fix test

* 491-delete-participant (#499)

* Update routes, docs and migrations

* Squash migration, update docs

* Fix tests

* Dev deployments (#477)

* main and dev builds

* rename to build.yml

* adds workflow dispatch for testing

* test

* test in branch

* remove if

* make changes to clients to test deployments

* Test using reusable workflow

* test again

* Try with deploy workflow

* try again after allowing access to repos in dsp

* inherit secrets

* adds ctrl-version

* test changes

* fix

* test

* test

* fix

* try with after build then push.

* try with after build then push

* revert dev test

* cleanup

* only run on workflow_run on main and dev

* run on successful pass only

* allow check node version

* fix type

* Make changes to test dev deployment workflow

* test

* test

* fix aeskey data check

* adds PRISMA_FIELD_ENCRYPTION_KEY to cron job

* test with updated flow

* use beta

* add --quiet

* push to correct ctrl helm chart dir

* setup gcp SDK

* rename the restart step

* fix the check

* revert dev deployment changes

* update to helm

* build images before helm

* echo statements

* test with == true

* only work on dev

* Started Enable merge profiles

* Started Enable merge profiles

* Started Enable merge profiles

* Fix Vulnerabilities (2025-07-22) (#490)

* 446-ci-release (#459)

* Update Helm Chart Version

* fix aeskey issue

* 460-study-mgmt-page (#462)

* Started feature 460-study-mgmt-page

* Study description and logo

* Add tests

* Update seed.ts

* 496-gen-id (#502)

* Started feature 496-gen-id

* Add participant IDs

* Fix tests

* 130-tcs (#501)

* Started feature 130-tcs

* Terms and conditions box

* Fix tests

* 492-log-recalc (#500)

* Started feature 492-log-recalc

* Log answer recalc

* Fix test

* 493-otp (#494)

* Started feature 493-otp

* OTP and password retries

* complete unfinished test

* Fix test

* 487-multi-study-families (#488)

* Started feature 487-multi-study-families

* Update family editing

* Add tests

* docs, fix test

* 491-delete-participant (#499)

* Update routes, docs and migrations

* Squash migration, update docs

* Fix tests

* update multer

* update axios to 1.11.0

* update vulnerability path

---------


* add debug statement for testing

* use the workflow run head branch to check if pushed to dev instead

* Update helm (#515)

* Update helm

* Helm updates

* fix: Docker image tag should check workflow_run

* terraform changes

* Terraform changes

* Fix Dev Deployments (#517)

* remove need to check helm changes

* change helm path in workflow

* checkout whatever branch the workflow run starts from.

* Fix types

* Fix typo

* Update terraform

* Update terraform

* Address review comments

* remove .env change

* Update

* Address PR comments

* Address PR comments

* Fix guardian count check

* remove logs

* Reset state on logout

* Fix lint

* Fix answer recalc bug

---------

+++ application/backend/src/controllers/StudiesController.ts
@@ -13,6 +13,8 @@ import {
   Middlewares,
   Controller,
   Security,
+  UploadedFile,
+  NoSecurity,
 } from 'tsoa'
 import logger from 'common/src/logger'
 import type {
@@ -24,13 +26,16 @@ import type {
 } from 'common/types/api/studies'
 import prisma from '../PrismaClient'
 import { Study } from '@prisma/client'
-import { NotFoundError } from '../middlewares/ErrorHandler'
+import { NotFoundError, UnprocessableError } from '../middlewares/ErrorHandler'
 import {
   ValidateErrorResponse,
   NotFoundErrorResponse,
   UnauthorizedErrorResponse,
 } from 'common/types/api/errors'
 import { auditLog } from '../middlewares/AuditLog'
+import sharp from 'sharp'
+import { Readable } from 'stream'
+import { SettingsController } from './SettingsController'
 
 @Route('studies')
 @Tags('Studies')
@@ -50,8 +55,8 @@ export class StudiesController extends Controller {
    */
   @Get('/')
   public async getAllStudies(): Promise<GetAllStudiesResponse> {
-    const studies: Study[] = await this.studyRepo.findMany({})
-    const responseData = { data: studies }
+    const studies: Study[] = await this.studyRepo.findMany({ orderBy: { id: 'asc' } })
+    const responseData = { data: studies.map((val) => ({ ...val, logo: Boolean(val.logo) })) }
     logger.info({ ...responseData })
     return responseData
   }
@@ -69,16 +74,14 @@ export class StudiesController extends Controller {
       where: { userId: request.user.userId },
     })
 
-    const studies: Study[] = await this.studyRepo.findMany({
-      where: {
-        profiles: {
-          some: {
-            participantProfileId: participantProfile.id,
-          },
-        },
-      },
+    const studies = await this.studyParticipantRepo.findMany({
+      where: { participantProfileId: participantProfile.id },
+      select: { study: true },
     })
-    const responseData = { data: studies }
+
+    const responseData = {
+      data: studies.map((val) => ({ ...val.study, logo: Boolean(val.study.logo) })),
+    }
     logger.info({ ...responseData })
     return responseData
   }
@@ -110,20 +113,17 @@ export class StudiesController extends Controller {
   @SuccessResponse('201', 'Created')
   @Response<ValidateErrorResponse>('422', 'Validation Failed')
   public async createStudy(@Body() bodyRequest: CreateStudyRequest): Promise<CreateStudyResponse> {
-    try {
-      const newStudy = await this.studyRepo.create({
-        data: { name: bodyRequest.name }, // Note: Study also has email invite info, but this is set via UI
-      })
-      const responseData = {
-        id: newStudy.id,
-      }
-      logger.info({ ...responseData })
-      return responseData
-    } catch (err) {
-      const errorMessage: string = 'Error creating study'
-      logger.error({ errorMessage, err })
-      throw new Error(errorMessage)
+    if ((await this.studyRepo.count({ where: { name: bodyRequest.name, deleted: false } })) > 0) {
+      throw new Error('Study with that name already exists')
     }
+    const newStudy = await this.studyRepo.create({
+      data: { name: bodyRequest.name }, // Note: Study also has email invite info, but this is set via UI
+    })
+    const responseData = {
+      id: newStudy.id,
+    }
+    logger.info({ ...responseData })
+    return responseData
   }
 
   /**
@@ -135,16 +135,41 @@ export class StudiesController extends Controller {
   @Response<NotFoundErrorResponse>('404', 'Not Found')
   @Response<ValidateErrorResponse>('422', 'Validation Failed')
   public async updateStudy(@Path() studyId: number, @Body() bodyRequest: UpdateStudyRequest) {
-    try {
-      await this.studyRepo.update({
-        where: { id: studyId },
-        data: bodyRequest,
-      })
-    } catch (err) {
-      const errorMessage: string = `Study with ID: ${studyId} not found`
-      logger.error({ errorMessage, err })
-      throw new NotFoundError(errorMessage)
+    if (
+      bodyRequest.name &&
+      (await this.studyRepo.count({ where: { name: bodyRequest.name, deleted: false } })) > 0
+    ) {
+      throw new Error('Study with that name already exists')
     }
+
+    await this.studyRepo.update({
+      where: { id: studyId },
+      data: bodyRequest,
+    })
+  }
+
+  @Post('/{studyId}/logo')
+  @Security('jwt', ['OrganisationAdmin'])
+  @Response<ValidateErrorResponse>('422', 'Validation Failed')
+  public async uploadLogo(@UploadedFile() file: Express.Multer.File, @Path() studyId: number) {
+    const buffer = await sharp(file.buffer).resize(200).png().toBuffer()
+    await prisma.study.update({ where: { id: studyId }, data: { logo: buffer } })
+  }
+
+  @Get('/{studyId}/logo')
+  @NoSecurity()
+  @Response<ValidateErrorResponse>('422', 'Validation Failed')
+  public async getLogo(@Path() studyId: number): Promise<Readable> {
+    const study = await prisma.study.findFirstOrThrow({
+      where: { id: studyId },
+      select: { logo: true },
+    })
+
+    if (!study.logo) {
+      return new SettingsController().getLogo()
+    }
+
+    return Readable.from(study.logo as Buffer)
   }
 
   /**
@@ -155,6 +180,11 @@ export class StudiesController extends Controller {
   @Delete('/{studyId}')
   @Response<NotFoundErrorResponse>('404', 'Not Found')
   public async deleteStudy(@Path() studyId: number) {
+    const studyCount = await this.studyRepo.count({})
+    if (studyCount == 1) {
+      throw new UnprocessableError('Must have at least one study')
+    }
+
     try {
       await this.studyRepo.delete({
         where: { id: studyId },
